Random List, Dictionary, String Generator
I am new to Python and required practice on Lists, Dictionaries, and Strings. I was slowed by constantly having to either find or create any of the prior items over and over, so I created this program from scratch with the meek knowledge I had available. 
Place this program at the top of your workbook(i.e. Jupyter). 
Input A in the first prompt if you want all ascii characters
Hit Enter for all letters and number
Then input the number of sequences you want in the item
Then input the length you want of each sequence
use the Variable zz for a list
use the Variable vv for a string
use the Variable ee for a dictionary
I hope you find this program as useful as I. By the way it is also great for generating passwords ;-)
Thanks,
Ralph Turchiano